==> R/01_load.qmd <==
---
title: "01_load"
format:
  html:
    self-contained: true
editor: visual
---

## Create folders to store data ("\_raw/", "data/")

```{r}
library(purrr)

folders <- c("../_raw/", "../data/")
folders |>
  walk(~ if (!dir.exists(.x)) dir.create(.x, recursive = TRUE))
```

## Data Retrieval from Online GitHub Repository

```{r}

url_ri_metadata = "https://raw.githubusercontent.com/msysbio/BTRI_mucin_study_data/refs/heads/main/01_RI_WC_mucin_metadata_all_RNA.csv"

url_ri_expression_data = "https://raw.githubusercontent.com/msysbio/BTRI_mucin_study_data/refs/heads/main/01_RI_WC_mucin_counts_all_RNA.csv"

url_ri_annotation = "https://raw.githubusercontent.com/msysbio/BTRI_mucin_study_data/refs/heads/main/Roseburia_Intestalis/Roseburia_intestinalis_annotation_EGGNOGG.csv"

url_ri_fermentation_data = "https://raw.githubusercontent.com/msysbio/BTRI_mucin_study_data/refs/heads/main/Fermentation%20data.csv"


ri_metadata = read.csv(url(url_ri_metadata))
ri_expression_data = read.csv(url(url_ri_expression_data))
ri_annotation = read.csv(url(url_ri_annotation))
ri_fermentation_data = read.csv(url(url_ri_fermentation_data))

write.csv(ri_metadata,"../_raw/01_metadata_data_load.csv")
write.csv(ri_expression_data,"../_raw/01_expression_data_load.csv")
write.csv(ri_annotation,"../_raw/01_annotation_data_load.csv")
write.csv(ri_fermentation_data,"../_raw/01_fermentation_data_load.csv")
```
