==> R/02_clean.qmd <==
---
title: "02_clean"
format:
  html:
    self-contained: true
---

# clean data

```{r}
#| message: false
library(tidyverse)
library(knitr)
```

## join expression and meta data

```{r}
ri_expression_data <- read.csv("../_raw/01_expression_data_load.csv")
ri_metadata <- read.csv("../_raw/01_metadata_data_load.csv")

ri_expression_data |> 
  pivot_longer(cols = -X,
               names_to = "sample",
               values_to = "expression") |> 
  inner_join(ri_metadata, by = c("sample" = "GC.code" )) |> 
  filter(Exp == "RI_batch_WC") |> 
  select(X.x,expression,sample) |> 
  pivot_wider(names_from = sample,
              values_from = expression) |> 
  write_csv("../data/02_expression_data_clean.csv")

knitr::kable(head(ri_expression_data))
```

## Clean annotation data

```{r}
ri_annotation <- read.csv("../_raw/01_annotation_data_load.csv")
ri_annotation |>
  select(gene_ids, ko) |>
  mutate(
    ko = ifelse(ko == "-", NA, ko)
  ) |>
  drop_na(ko) |>
  write_csv("../data/02_annotation_data_clean.csv")

knitr::kable(head(ri_annotation))
```

## Clean fermentation data

```{r}
ri_fermentation <- read.csv("../_raw/01_fermentation_data_load.csv")
ri_fermentation |>
  filter(Exp == "RI_batch_WC1") |>
  select( "Rowname" | (starts_with("T") & ends_with("h")) ) |>
  select(where(~!all(is.na(.x)) ) ) |>
  write_csv("../data/02_fermentation_data_clean.csv")

knitr::kable(head(ri_fermentation))
```
