==> R/03_augment.qmd <==
---
title: "03_augment"
format:
  html:
    self-contained: true
---
```{r}
#| message: false
library(tidyverse)
library(knitr)
``` 

```{r}
#| warning: false
ri_metadata <- read.csv("../_raw/01_metadata_data_load.csv")
ri_expression_clean <- read.csv("../data/02_expression_data_clean.csv")
ri_expression_clean |> 
  pivot_longer(cols = -X.x,
               names_to = "sample",
               values_to = "expression") |> 
  inner_join(ri_metadata, by = c("sample" = "GC.code" )) |> 
  filter(Exp == "RI_batch_WC") |> 
  select(X,expression,sample) |> 
  pivot_wider(names_from = sample,
              values_from = expression) |> 
  write_csv("../data/03_expression_data_augment.csv")

knitr::kable(head(ri_expression_clean))
```

```{r}
ri_annotation_clean <- read.csv("../data/02_annotation_data_clean.csv")
ri_annotation_clean |>
  separate_rows(ko, sep = ",") |>
  write_csv("../data/03_annotation_data_augment.csv")

knitr::kable(head(ri_annotation_clean))
```
