==> R/07_analyse_3.qmd <==
---
title: "05_analyse_03"
format:
  html:
    self-contained: true
---

## Load Data

```{R}
#Load libraries
library("tidyverse")
library("ggplot2")

ri_fermentation_data <- read.csv("../_raw/01_fermentation_data_load.csv")

#Subset the columns used for the plots
subset_cols <- c("T0h", "T4h", "T8h", "T12h", "T16h", "T24h", "T28h", "T32h", "T48h", "T60h", "T72h", "T96h", "T120h")

#First plot
cell_counts_map <- list(
  "Total_count"  = c(139, 140, 141, 274, 275),
  "Active_count" = c(142, 143, 144, 276, 277)
)

#Second plot
metabolite_map <- list(
  "Glucose"  = c(145, 146, 147, 278, 279),
  "Pyruvate" = c(151, 152, 153, 282, 283),
  "Lactate"  = c(154, 155, 156, 284, 285),
  "Formate"  = c(157, 158, 159, 286, 287),
  "Acetate"  = c(160, 161, 162, 288, 289),
  "Butyrate" = c(163, 164, 165, 290, 291),
  "pH"       = c(136, 137, 138, 272, 273)
)
```

## Augment data

```{R}

#Function for augmenting data
process_metabolite <- function(rows, name, data) {
  
  #Subset and Summarise the metrics of interest
  data[rows, subset_cols, drop=FALSE] |>
    summarise(across(everything(), list(mean = mean, sd = sd))) |>
    
  #Pivot that uses either hours, or the metrics as names
    pivot_longer(
      cols = everything(), 
      names_to = c("Hours", ".value"), 
      names_sep = "_"
    ) |>
    mutate(group = name)
}
# First Plot
plot_df <- imap_dfr(cell_counts_map, ~ process_metabolite(.x, .y, ri_fermentation_data)) |>
  mutate(Hours = factor(Hours, levels = subset_cols))

#Second Plot
plot_mm_df <- imap_dfr(metabolite_map, ~ process_metabolite(.x, .y, ri_fermentation_data)) |> mutate(Hours = factor(Hours, levels = subset_cols)) |> 
  mutate(mean = ifelse(group == "pH", (mean - 5) * 6, mean))
```

## Analyse data

```{R}
#First plot

ggplot(plot_df, aes(x = Hours, y = mean, group = group, linetype = group, color = group))+
  geom_line(linewidth = 1)+
  geom_point()+
  geom_errorbar(aes(ymin = mean-sd, ymax = mean+sd), width = 0.2)+
  scale_color_manual(values = c("Total_count" = "#FC8D62", "Active_count" = "#66C2A5"))+
  scale_linetype_manual(values = c("Total_count" = "solid", "Active_count" = "dashed"))+
  labs(title = "Mean Growth Curve for RI", x = "Hours", y = "Cell count", 
       color = "Group", linetype = "Group")+
  theme_minimal()
```

```{R}
#Second plot

ggplot(plot_mm_df, aes(x = Hours, y = mean, group = group, linetype = group, color = group))+
  geom_line(size = 1)+
  geom_point(size = 2)+
  scale_y_continuous(
    name = "Concentration (mM)",
    limits = c(0,12),
    sec.axis = sec_axis(~ (. / 6) + 5, name = "pH", breaks = seq(5,7,0.5))
  )+
  labs(title = "Mean Growth Curve for RI", x = "Hours", y = "mM", 
       color = "Group", linetype = "Group")+
  theme_minimal()
```
